fix(app/signIn/page):se cambia la page a jsx y se anexa el enviado de mail
