+++ tsconfig.json
@@ -28,7 +28,7 @@
     ".next/types/**/*.ts",
     "**/*.ts",
     "**/*.tsx"
-  ],
+, "app/signin/page.jsx"  ],
   "exclude": [
     "node_modules"
   ]
